test-requirements.txt
GitPython>=1.0.1
Flask>=0.10.1
PyYAML>=3.11
tox
pep8
pytest
